for uknown requests send 404 status instead of redirecting to a 404 content page with a 200 status
