+++ api/controllers/main.js
@@ -1,30 +1,6 @@
 const config = require('../../config');
 const SiteWideErrorLoader = require('../services/SiteWideErrorLoader');
-const { loadAssetManifest, getSiteDisplayEnv, shouldIncludeTracking } = require('../utils');
-const Features = require('../features');
-
-const webpackAssets = loadAssetManifest();
-
-function defaultContext(req, res) {
-  const messages = {
-    errors: req.flash('error'),
-  };
-
-  const context = {
-    isAuthenticated: false,
-    messages,
-    shouldIncludeTracking: shouldIncludeTracking(),
-    siteDisplayEnv: getSiteDisplayEnv(),
-    homepageUrl: config.app.homepageUrl,
-    webpackAssets,
-    authGithub: Features.enabled(Features.Flags.FEATURE_AUTH_GITHUB),
-    authUAA: Features.enabled(Features.Flags.FEATURE_AUTH_UAA),
-    hasUAAIdentity: false,
-    nonce: res.locals.cspNonce,
-  };
-
-  return context;
-}
+const { defaultContext } = require('../utils');
 
 function alertGithubAuthDeprecation(hasUAAIdentity, context) {
   if (!hasUAAIdentity && context.authUAA) {
@@ -87,15 +63,5 @@ module.exports = {
 
     // send the "deny all" robots.txt content
     return res.send(DENY_ALL_CONTENT);
-  },
-
-  notFound(req, res) {
-    const context = defaultContext(req, res);
-    if (req.session.authenticated) {
-      context.isAuthenticated = true;
-      context.username = req.user.username;
-    }
-
-    res.render('404.njk', context);
   },
 };
